+++ modules/@apostrophecms/page/index.js
@@ -254,14 +254,13 @@ module.exports = {
       // This call is atomic with respect to other REST write operations on pages.
       async post(req) {
         self.publicApiCheck(req);
-        req.body._position = req.body._position || 'lastChild';
         let targetId = self.apos.launder.string(req.body._targetId);
-        let position = self.apos.launder.string(req.body._position);
+        let position = self.apos.launder.string(req.body._position || 'lastChild');
         // Here we have to normalize before calling insert because we
         // need the parent page to call newChild(). insert calls again but
         // sees there's no work to be done, so no performance hit
         const normalized = await self.getTargetIdAndPosition(req, null, targetId, position);
-        targetId = normalized.targetId;
+        targetId = normalized.targetId || '_home';
         position = normalized.position;
         const copyingId = self.apos.launder.id(req.body._copyingId);
         const input = _.omit(req.body, '_targetId', '_position', '_copyingId');
@@ -273,7 +272,7 @@ module.exports = {
         if (req.body._newInstance) {
           // If we're looking for a fresh page instance and aren't saving yet,
           // simply get a new page doc and return it
-          const parentPage = await self.findForEditing(req, { _id: targetId })
+          const parentPage = await self.findForEditing(req, self.getIdCriteria(targetId))
             .permission('edit', '@apostrophecms/any-page-type').toObject();
           const newChild = self.newChild(parentPage);
           newChild._previewable = true;
@@ -281,7 +280,12 @@ module.exports = {
         }
 
         return self.withLock(req, async () => {
-          const targetPage = await self.findForEditing(req, targetId ? self.getIdCriteria(targetId) : { level: 0 }).ancestors(true).permission('edit').toObject();
+          const targetPage = await self
+            .findForEditing(req, self.getIdCriteria(targetId))
+            .ancestors(true)
+            .permission('edit')
+            .toObject();
+
           if (!targetPage) {
             throw self.apos.error('notfound');
           }