+++ modules/@apostrophecms/schema/index.js
@@ -523,15 +523,29 @@ module.exports = {
         const errorsList = [];
 
         for (const error of errors) {
-          if ((error.name === 'required' || error.name === 'mandatory')) {
+          if (error.path) {
             // `self.isVisible` will only throw for required fields that have
             // an external condition containing an unknown module or method:
             const isVisible = await self.isVisible(req, schema, destination, error.path);
 
             if (!isVisible) {
-              // It is not reasonable to enforce required for
-              // fields hidden via conditional fields
-              continue;
+              // It is not reasonable to enforce required,
+              // min, max or anything else for fields
+              // hidden via "if" as the user cannot correct it
+              // and it will not be used. If the user changes
+              // the conditional field later then they won't
+              // be able to save until the erroneous field
+              // is corrected
+              const name = error.path;
+              const field = schema.find(field => field.name === name);
+              if (field) {
+                // To protect against security issues, an invalid value
+                // for a field that is not visible should be quietly discarded.
+                // We only worry about this if the value is not valid, as otherwise
+                // it's a kindness to save the work so the user can toggle back to it
+                destination[field.name] = (field.def != undefined) ? field.def : self.fieldTypes[field.type]?.def;
+                continue;
+              }
             }
           }